refactor(list.c): make types match
Also allows the i to handle bigger cases

+++ src/list.c
@@ -137,7 +137,7 @@ CommandList getList(void) {
         arrayList[i] = "";
     }
 
-    int i = 0;
+    intmax_t i = 0;
     for (char* item = strtok(items, "\n"); item; item = strtok(NULL, "\n")) {
         arrayList[i++] = item;
     }
